[TLC 221] fixed header and browse_global menu layout and behavior on mobile devices (xs, sm).
For GI there is only one menu-point in the public menu, "Discover". Accordingly this should be treated differently than the default (with statistics menu for public). Now the section opens when the menu is displayed in a screen pulldown on mobile devices. Should be changed accordingly if more menu sections need to be present.