Integrated home-news.component and GiSearchForm into GI theme
The GI search form resides in a shared folder. It needs to be declared somewhere within the declaration hierarchy, so that the other components can access it.

Because of the lack of another obvious place for the time being it is declared within the eager theme module, however this does not seem to be ideal.

Declaring it within the lazy theme module also doesn't cut it, because this module says explicitly it should NEVER be referenced and it isn't. It makes sure that the component is compiled, but it still is not within the scope of the other components.